Convert to Composer
* convert folder layout to PSR-4 standard
* convert code to PSR-2 standard
* remove PEAR related files and documentation
* add composer.json
* update Yahoo weather API integration
