+++ .gitignore
@@ -1,6 +1,7 @@
 .build
 .idea
 .tmp
-dist
+composer.lock
+nbproject
 review
 vendor
