+++ LICENSE.txt
@@ -1,7 +1,7 @@
 New BSD License
 ===============
 
-Copyright (c) 2014, random-host.com
+Copyright (c) 2016, random-host.com
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
