--- build.properties
@@ -1,14 +1,0 @@
-project.name=PHP_Weather
-project.channel=pear.random-host.com
-project.majorVersion=0
-project.minorVersion=4
-project.patchLevel=2
-project.stability=beta
-project.snapshot=true
-
-checkstyle.standard=PEAR
-
-component.type=php-library
-component.version=12
-
-pear.local=/var/www/${project.channel}
