+++ composer.json
@@ -1,0 +1,34 @@
+{
+    "name": "randomhost/weather",
+    "type": "library",
+    "description": "Encapsulates functions for easy retrieval of weather data from the Yahoo Weather API.",
+    "keywords": [
+        "Weather"
+    ],
+    "homepage": "https://www.random-host.com",
+    "license": "New-BSD",
+    "authors": [
+        {
+            "name": "Ch'Ih-Yu",
+            "email": "chi-yu@web.de"
+        }
+    ],
+    "config": {
+        "platform": {
+            "php": "5.3.3"
+        }
+    },
+    "require": {
+        "php": ">=5.3.0"
+    },
+    "require-dev": {
+        "behat/behat": "~2.0",
+        "phpunit/phpunit": "~4.0",
+        "phpspec/phpspec": "~2.4"
+    },
+    "autoload": {
+        "psr-4": {
+            "randomhost\\Weather\\": "src/php/"
+        }
+    }
+}
