--- package.xml
@@ -1,182 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<package packagerversion="1.9.1" version="2.0"
-         xmlns="http://pear.php.net/dtd/package-2.0"
-         xmlns:tasks="http://pear.php.net/dtd/tasks-1.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://pear.php.net/dtd/tasks-1.0
-    http://pear.php.net/dtd/tasks-1.0.xsd
-    http://pear.php.net/dtd/package-2.0
-    http://pear.php.net/dtd/package-2.0.xsd">
-    <name>${project.name}</name>
-    <channel>${project.channel}</channel>
-    <summary>
-        Encapsulates SimpleXML functions for easy retrieval of weather data
-        from the Yahoo Weather API.
-    </summary>
-    <description>
-        This package encapsulates SimpleXML functions for easy retrieval of
-        weather data from the Yahoo Weather API. It was created as part of the
-        PHP_Webcam_Overlay package but is released as a separate component so it
-        can be used in other packages.
-
-        Because it was created as a dependency of the PHP_Webcam_Overlay package,
-        it does only support a subset of the available weather data.
-    </description>
-    <lead>
-        <name>Ch'Ih-Yu</name>
-        <user>chi-yu</user>
-        <email>chi-yu@web.de</email>
-        <active>yes</active>
-    </lead>
-    <date>${build.date}</date>
-    <time>${build.time}</time>
-    <version>
-        <release>${project.version}</release>
-        <api>${project.majorVersion}.${project.minorVersion}</api>
-    </version>
-    <stability>
-        <release>${project.stability}</release>
-        <api>beta</api>
-    </stability>
-    <license>New-BSD</license>
-    <notes>
-        See https://github.com/Random-Host/${project.name}/ for more details.
-    </notes>
-    <contents>
-        <dir baseinstalldir="/" name="/">
-            ${contents}
-        </dir>
-    </contents>
-    <dependencies>
-        <required>
-            <php>
-                <min>5.3.0</min>
-            </php>
-            <pearinstaller>
-                <min>1.9.4</min>
-            </pearinstaller>
-            <package>
-                <name>Autoloader</name>
-                <channel>pear.phix-project.org</channel>
-                <min>3.0.0</min>
-                <max>3.999.9999</max>
-            </package>
-        </required>
-    </dependencies>
-    <phprelease/>
-    <changelog>
-        <release>
-            <version>
-                <release>0.4.2</release>
-                <api>0.4</api>
-            </version>
-            <stability>
-                <release>beta</release>
-                <api>beta</api>
-            </stability>
-            <date>2014-03-17</date>
-            <license>New-BSD</license>
-            <notes>
-                Fixed incorrect inline documentation and updated the README.
-                This version does not include any code changes.
-
-                See https://github.com/Random-Host/${project.name}/ for more
-                details.
-            </notes>
-        </release>
-        <release>
-            <version>
-                <release>0.4.1</release>
-                <api>0.4</api>
-            </version>
-            <stability>
-                <release>beta</release>
-                <api>beta</api>
-            </stability>
-            <date>2014-03-14</date>
-            <license>New-BSD</license>
-            <notes>
-                Added the previously forgotten PEAR changelog entries.
-                This version does not include any code changes.
-
-                See https://github.com/Random-Host/${project.name}/ for more
-                details.
-            </notes>
-        </release>
-        <release>
-            <version>
-                <release>0.4.0</release>
-                <api>0.4</api>
-            </version>
-            <stability>
-                <release>beta</release>
-                <api>beta</api>
-            </stability>
-            <date>2014-03-14</date>
-            <license>New-BSD</license>
-            <notes>
-                Added Forecast class for retrieving weather forecasts.
-
-                See https://github.com/Random-Host/${project.name}/ for more
-                details.
-            </notes>
-        </release>
-        <release>
-            <version>
-                <release>0.3.0</release>
-                <api>0.3</api>
-            </version>
-            <stability>
-                <release>beta</release>
-                <api>beta</api>
-            </stability>
-            <date>2014-03-07</date>
-            <license>New-BSD</license>
-            <notes>
-                Unit test for all classes were added.
-                Some classes received minor edits to be better testable.
-
-                See https://github.com/Random-Host/${project.name}/ for more
-                details.
-            </notes>
-        </release>
-        <release>
-            <version>
-                <release>0.2.0</release>
-                <api>0.2</api>
-            </version>
-            <stability>
-                <release>alpha</release>
-                <api>alpha</api>
-            </stability>
-            <date>2014-03-03</date>
-            <license>New-BSD</license>
-            <notes>
-                Moved data into data access objects for easier and more modern
-                handling.
-
-                See https://github.com/Random-Host/${project.name}/ for more
-                details.
-            </notes>
-        </release>
-        <release>
-            <version>
-                <release>0.1.0</release>
-                <api>0.1</api>
-            </version>
-            <stability>
-                <release>alpha</release>
-                <api>alpha</api>
-            </stability>
-            <date>2014-03-01</date>
-            <license>New-BSD</license>
-            <notes>
-                Initial PEAR packaged version
-
-                See https://github.com/Random-Host/${project.name}/ for more
-                details.
-            </notes>
-        </release>
-    </changelog>
-</package>
-<!-- vim: set tabstop=4 shiftwidth=4 expandtab:4 -->
