+++ src/data/test/feedInvalid.json
@@ -1,0 +1,1 @@
+This is not JSON
