--- src/tests/unit-tests/php/testdata/feedNamespaceError.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
-<rss version="2.0">
-    <channel>
-        <title>Yahoo! Weather - Cologne, DE</title>
-        <link>http://us.rd.yahoo.com/dailynews/rss/weather/Cologne__DE/*http://weather.yahoo.com/forecast/GMXX0018_f.html</link>
-        <description>Yahoo! Weather for Cologne, DE</description>
-        <language>en-us</language>
-        <lastBuildDate>Fri, 07 Mar 2014 8:48 pm CET</lastBuildDate>
-        <ttl>60</ttl>
-        <yweather:location city="Cologne" region="NW" country="Germany"/>
-        <yweather:units temperature="F" distance="mi" pressure="in" speed="mph"/>
-        <yweather:wind chill="48" direction="340" speed="5"/>
-        <yweather:atmosphere humidity="62" visibility="6.21" pressure="30.33" rising="0"/>
-        <yweather:astronomy sunrise="7:03 am" sunset="6:21 pm"/>
-        <image>
-            <title>Yahoo! Weather</title>
-            <width>142</width>
-            <height>18</height>
-            <link>http://weather.yahoo.com</link>
-            <url>http://l.yimg.com/a/i/brand/purplelogo//uh/us/news-wea.gif
-            </url>
-        </image>
-        <item>
-            <title>Conditions for Cologne, DE at 8:48 pm CET</title>
-            <geo:lat>50.94</geo:lat>
-            <geo:long>6.96</geo:long>
-            <link>http://us.rd.yahoo.com/dailynews/rss/weather/Cologne__DE/*http://weather.yahoo.com/forecast/GMXX0018_f.html</link>
-            <pubDate>Fri, 07 Mar 2014 8:48 pm CET</pubDate>
-            <yweather:condition text="Fair" code="33" temp="50" date="Fri, 07 Mar 2014 8:48 pm CET"/>
-            <description><![CDATA[
-<img src="http://l.yimg.com/a/i/us/we/52/33.gif"/><br />
-<b>Current Conditions:</b><br />
-Fair, 50 F<BR />
-<BR /><b>Forecast:</b><BR />
-Fri - Cloudy. High: 58 Low: 41<br />
-Sat - Partly Cloudy. High: 61 Low: 42<br />
-Sun - Partly Cloudy. High: 67 Low: 38<br />
-Mon - Sunny. High: 64 Low: 38<br />
-Tue - Sunny. High: 57 Low: 34<br />
-<br />
-<a href="http://us.rd.yahoo.com/dailynews/rss/weather/Cologne__DE/*http://weather.yahoo.com/forecast/GMXX0018_f.html">Full Forecast at Yahoo! Weather</a><BR/><BR/>
-(provided by <a href="http://www.weather.com" >The Weather Channel</a>)<br/>
-]]></description>
-            <yweather:forecast day="Fri" date="7 Mar 2014" low="41" high="58" text="Cloudy" code="26"/>
-            <yweather:forecast day="Sat" date="8 Mar 2014" low="42" high="61" text="Partly Cloudy" code="30"/>
-            <yweather:forecast day="Sun" date="9 Mar 2014" low="38" high="67" text="Partly Cloudy" code="30"/>
-            <yweather:forecast day="Mon" date="10 Mar 2014" low="38" high="64" text="Sunny" code="32"/>
-            <yweather:forecast day="Tue" date="11 Mar 2014" low="34" high="57" text="Sunny" code="32"/>
-            <guid isPermaLink="false">GMXX0018_2014_03_11_7_00_CET</guid>
-        </item>
-    </channel>
-</rss>
